traffic_analysis/requirements.txt
gdown
inference==0.9.17
supervision>=0.20.0
tqdm
ultralytics<=8.3.40
fastapi
uvicorn
psycopg2-binary
celery
redis
python-multipart